fix: monitor use u64

+++ src/data/regular.rs
@@ -94,7 +94,7 @@ pub enum Transforms {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Monitor {
     /// The monitor id
-    pub id: i16,
+    pub id: u64,
     /// The monitor's name
     pub name: String,
     /// The monitor's description
@@ -217,7 +217,7 @@ pub struct Client {
     #[serde(rename = "fullscreenMode")]
     pub fullscreen_mode: i8,
     /// The monitor the window is on
-    pub monitor: i16,
+    pub monitor: u64,
     /// The initial window class
     #[serde(rename = "initialClass")]
     pub initial_class: String,
